svg-edit: Try to import jQuery
To import jQuery as a library the code has to be condensed to _one_
evaluateJavaScript - otherwise the jQuery object wouldn't be available
inside editsvg.js (don't know why this doesn't work).

However, as soon as jquery is included there will be a unspecified
"Script error" somewhere in editsvg.js.
I don't know what causes this but it would be worth checking jQuery out.